backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3
numpy-financial==1.0.0
pdfplumber==0.9.0
pgvector==0.2.3
openai==1.28.0
langchain==0.1.0
redis==4.5.5
celery==5.3.1
python-multipart==0.0.6
faiss-cpu==1.7.4
httpx==0.27.2